test(mfa): split the graph-traversal suites into a dedicated subfolder
One 213-line file mixed chart-only suites with the real-UI walk, so the
four module-level jest mocks also applied to suites that never render.
The suites now live in machine/graphTraversal/, split by guarantee:
everyStateReachable.test.ts analyzes the chart alone, that is
reachability and journey endpoints; coverageStaysComplete.test.ts
guards the generated walk coverage without rendering; and
viewMatchesMachine.test.tsx keeps the mocks, the executors, the walk,
and the testConfig completeness guards next to the config they pin.
The subfolder keeps the model-based family apart from the plain machine
tests that will land in machine/ later.
