+++ tests/unit/components/MultifactorAuthentication/machine/graphTraversal/everyStateReachable.test.ts
@@ -1,0 +1,33 @@
+import {getDrivingJourneyPaths, getMfaShortestPaths} from 'tests/utils/mfa/flowPaths';
+import getSettleableLeafStates from 'tests/utils/mfa/settleableLeafStates';
+import {matchesState} from 'xstate';
+import mfaMachine from '@components/MultifactorAuthentication/machine/mfaMachine';
+
+// Both suites in this file are generated by traversing the mfaMachine graph automatically instead of
+// enumerating cases by hand, and both analyze the chart alone; nothing renders. `getShortestPaths`
+// derives the coverage from the chart, so a state added to the machine gains a spec here without edits.
+// `settleableLeafStates.ts` defines which states count as settleable. The sibling suites under
+// `graphTraversal/` walk the same generated paths through the real modal UI and guard that coverage.
+
+// Reachability is asserted on the unfiltered shortest paths, so it also covers routes the UI walk in
+// `viewMatchesMachine.test.tsx` cannot drive (a delayed transition, for example). A failure here means
+// the chart itself lost the state, not that the view harness broke.
+describe('every settleable MFA state is reachable in the machine chart', () => {
+    const reachableSnapshots = getMfaShortestPaths().map((path) => path.state);
+
+    it.each(getSettleableLeafStates(mfaMachine.root))('$description can be reached from the initial state', ({description}) => {
+        expect(reachableSnapshots.some((snapshot) => matchesState(description, snapshot.value))).toBe(true);
+    });
+});
+
+// Every generated expectation in these suites follows the machine under test, so a retargeted transition
+// that keeps every state reachable regenerates matching expectations and passes. Each journey's pinned
+// endpoint is hand-written in `DRIVING_JOURNEYS`, so it fails in that case instead.
+describe('every driving journey ends in its pinned state', () => {
+    it.each(getDrivingJourneyPaths())('$description', ({paths, endState}) => {
+        expect(paths.length).toBeGreaterThan(0);
+        for (const path of paths) {
+            expect(matchesState(endState, path.state.value)).toBe(true);
+        }
+    });
+});
