+++ tests/unit/components/MultifactorAuthentication/machine/graphTraversal/viewMatchesMachine.test.tsx
@@ -1,11 +1,10 @@
 import {act, fireEvent, screen} from '@testing-library/react-native';
 import {describeTraversalEvent, isTestScenarioInitEvent} from 'tests/utils/mfa/flowFixtures';
-import getWalkedPaths, {getDrivingJourneyPaths, getMfaShortestPaths} from 'tests/utils/mfa/flowPaths';
+import getWalkedPaths from 'tests/utils/mfa/flowPaths';
 import {getMfaControls, renderMfaUi} from 'tests/utils/mfa/realUi/harness';
 import {pendingModalClose, resetMfaUiMocks} from 'tests/utils/mfa/realUi/mocks';
 import type * as MfaRealUiMocks from 'tests/utils/mfa/realUi/mocks';
 import getSettleableLeafStates from 'tests/utils/mfa/settleableLeafStates';
-import {getExercisedTransitionKeys, getInitEdgeLandings, getUiDrivableTransitions} from 'tests/utils/mfa/transitionCoverage';
 import waitForBatchedUpdatesWithAct from 'tests/utils/waitForBatchedUpdatesWithAct';
 import {matchesState} from 'xstate';
 import mfaMachine from '@components/MultifactorAuthentication/machine/mfaMachine';
@@ -14,12 +13,11 @@ import {handleInitialScreenLayout, mfaNavigationRef} from '@components/Multifact
 import CONST from '@src/CONST';
 import SCREENS from '@src/SCREENS';
 
-// Every suite in this file is generated by traversing the mfaMachine graph automatically instead of
-// enumerating cases by hand. `getShortestPaths` derives the coverage from the chart, so a state or
-// event added to the machine gains specs here without edits. The file layers the chart itself (every
-// settleable state is reachable; `settleableLeafStates.ts` defines which states count as settleable),
-// the pinned endpoints of the driving journeys, the contract between the machine and the rendered
-// modal (the walk), and guards proving that the generated coverage itself stays complete.
+// This file pins the contract between the machine and the rendered modal: it mounts the production
+// providers and navigator, drives each machine event as a real gesture, and asserts the UI markers of
+// the reached state at every step of every generated path. The paths come from `getWalkedPaths`, so a
+// state or event added to the machine gains steps here without edits. The chart-only suites live in
+// `everyStateReachable.test.ts`, and the guards on the generated coverage in `coverageStaysComplete.test.ts`.
 
 // This mock forces a wide layout so the navigator renders the backdrop used as the mounted marker.
 jest.mock('@hooks/useResponsiveLayout');
@@ -33,29 +31,6 @@ jest.mock('@components/MultifactorAuthentication/biometrics/useBiometrics', () =
 jest.mock('@components/MultifactorAuthentication/useSyncMfaModalNavigatorWithHistory', () => jest.requireActual<typeof MfaRealUiMocks>('tests/utils/mfa/realUi/mocks').syncHistoryMock());
 // This mock reuses the shared Navigation implementation and overrides the transition methods used by the MFA flow.
 jest.mock('@libs/Navigation/Navigation', () => jest.requireActual<typeof MfaRealUiMocks>('tests/utils/mfa/realUi/mocks').navigationMock());
-
-// Reachability is asserted on the unfiltered shortest paths, so it also covers routes the UI walk
-// below cannot drive (a delayed transition, for example). A failure here means the chart itself lost
-// the state, not that the view harness broke.
-describe('every settleable MFA state is reachable in the machine chart', () => {
-    const reachableSnapshots = getMfaShortestPaths().map((path) => path.state);
-
-    it.each(getSettleableLeafStates(mfaMachine.root))('$description can be reached from the initial state', ({description}) => {
-        expect(reachableSnapshots.some((snapshot) => matchesState(description, snapshot.value))).toBe(true);
-    });
-});
-
-// Every generated expectation in this file follows the machine under test, so a retargeted transition
-// that keeps every state reachable regenerates matching expectations and passes. Each journey's pinned
-// endpoint is hand-written in `DRIVING_JOURNEYS`, so it fails in that case instead.
-describe('every driving journey ends in its pinned state', () => {
-    it.each(getDrivingJourneyPaths())('$description', ({paths, endState}) => {
-        expect(paths.length).toBeGreaterThan(0);
-        for (const path of paths) {
-            expect(matchesState(endState, path.state.value)).toBe(true);
-        }
-    });
-});
 
 // These UI markers distinguish the closed, closing, and outcome states. `OutcomeScreenBase` identifies the
 // outcome screen, while the backdrop exists only when the MFA navigator is mounted. Every outcome screen
@@ -160,61 +135,22 @@ describe('the real MFA modal matches the machine at every step of every generate
     }
 });
 
-// A failure below usually means the walk stopped covering a state or transition, or the view config
-// drifted from the machine, not that the modal broke. Each failure names the missing piece.
-describe('the generated coverage itself stays complete', () => {
-    // Every settleable leaf must end a path that the UI walk drives, not merely appear mid-path.
-    describe('every settleable MFA state is reachable through the real UI', () => {
-        const walkedLeafValues = walkedPaths.map((path) => path.state.value);
+// TestModel runs only the state assertions whose keys match the reached state, so if no key matches a
+// state, the test passes without checking it. These guards fail in that case.
+//
+// Types cannot do this. TypeScript's inferred type does not record whether a state has an `always`
+// transition, so it cannot tell a real leaf from a pass-through state such as `{open: "preparing"}`. A
+// `Record<leaf, ...>` would then need an empty assertion for every pass-through state, and that empty
+// entry would make the check pass on its own once the state later loses its `always` and becomes settleable.
+describe('testConfig defines a UI assertion for every settleable state and for nothing else', () => {
+    const settleableLeafStates = getSettleableLeafStates(mfaMachine.root);
+    const configuredStateKeys = Object.keys(testConfig.states);
 
-        it.each(getSettleableLeafStates(mfaMachine.root))('$description can be reached through the real UI', ({description}) => {
-            expect(walkedLeafValues.some((reached) => matchesState(description, reached))).toBe(true);
-        });
+    it.each(settleableLeafStates)('$description has a UI assertion', ({description}) => {
+        expect(configuredStateKeys.some((key) => matchesState(key, description))).toBe(true);
     });
 
-    // Shortest paths keep a single incoming route per distinct state-and-context vertex, so a transition can
-    // silently drop out of the walk once another route to its target is shorter. A failure here has two
-    // possible causes. When a transition was added to the machine, the fix is an explicit journey in
-    // `DRIVING_JOURNEYS` that drives it. When the machine stopped resetting part of its context, the same
-    // state value splits into a second vertex whose transitions have no route, so compare the failing
-    // vertex's context with the intended one and fix the machine instead of adding a journey.
-    describe('every UI-drivable state-changing transition is exercised', () => {
-        const exercisedTransitionKeys = getExercisedTransitionKeys(walkedPaths);
-
-        it.each(getUiDrivableTransitions())('$description is exercised', ({key}) => {
-            expect(exercisedTransitionKeys.has(key)).toBe(true);
-        });
-    });
-
-    // The payload INIT fixture exists to give the payload flow its own context vertices. When the machine
-    // stops copying the payload into its context, every INIT fixture lands in the same vertex, the split
-    // disappears, and the explicit journeys keep the walk green. This guard pins the split itself.
-    describe('INIT fixtures produce distinct context vertices', () => {
-        it('keeps at least one landing vertex per distinct INIT fixture', () => {
-            const landings = getInitEdgeLandings();
-            const distinctFixtures = new Set(landings.map(({eventKey}) => eventKey));
-            const distinctLandings = new Set(landings.map(({landingKey}) => landingKey));
-            expect(distinctLandings.size).toBeGreaterThanOrEqual(distinctFixtures.size);
-        });
-    });
-
-    // TestModel runs only the state assertions whose keys match the reached state, so if no key matches a
-    // state, the test passes without checking it. These guards fail in that case.
-    //
-    // Types cannot do this. TypeScript's inferred type does not record whether a state has an `always`
-    // transition, so it cannot tell a real leaf from a pass-through state such as `{open: "preparing"}`. A
-    // `Record<leaf, ...>` would then need an empty assertion for every pass-through state, and that empty
-    // entry would make the check pass on its own once the state later loses its `always` and becomes settleable.
-    describe('testConfig defines a UI assertion for every settleable state and for nothing else', () => {
-        const settleableLeafStates = getSettleableLeafStates(mfaMachine.root);
-        const configuredStateKeys = Object.keys(testConfig.states);
-
-        it.each(settleableLeafStates)('$description has a UI assertion', ({description}) => {
-            expect(configuredStateKeys.some((key) => matchesState(key, description))).toBe(true);
-        });
-
-        it.each(configuredStateKeys.map((key) => ({key})))('$key matches a real settleable state', ({key}) => {
-            expect(settleableLeafStates.some((leaf) => matchesState(key, leaf.description))).toBe(true);
-        });
+    it.each(configuredStateKeys.map((key) => ({key})))('$key matches a real settleable state', ({key}) => {
+        expect(settleableLeafStates.some((leaf) => matchesState(key, leaf.description))).toBe(true);
     });
 });
